+++ src/components/Login/LoginForm.js
@@ -1,100 +1,58 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { connect as reduxConnect } from 'react-redux';
-import {
-  withRouter,
-  matchPath,
-} from 'react-router-dom';
+import { FormattedMessage } from 'react-intl';
+import { Field, Form } from 'react-final-form';
 
-import { ConnectContext } from '@folio/stripes-connect';
-import {
-  requestLogin,
-  requestSSOLogin,
-} from '../../loginServices';
-import { setAuthError } from '../../okapiActions';
+import { branding } from 'stripes-config';
 
-class LoginCtrl extends Component {
+import {
+  TextField,
+  Button,
+  Row,
+  Col,
+  Headline,
+} from '@folio/stripes-components';
+
+import SSOLogin from '../SSOLogin';
+import OrganizationLogo from '../OrganizationLogo';
+import AuthErrorsContainer from '../AuthErrorsContainer';
+import FieldLabel from '../CreateResetPassword/components/FieldLabel';
+
+import styles from './Login.css';
+
+class LoginForm extends Component {
   static propTypes = {
-    authFailure: PropTypes.arrayOf(PropTypes.object),
-    ssoEnabled: PropTypes.bool,
-    autoLogin: PropTypes.shape({
-      username: PropTypes.string.isRequired,
-      password: PropTypes.string.isRequired,
-    }),
-    clearAuthErrors: PropTypes.func.isRequired,
-    history: PropTypes.shape({
-      push: PropTypes.func.isRequired,
-    }).isRequired,
-    location: PropTypes.shape({
-      pathname: PropTypes.string.isRequired,
-    }).isRequired,
+    ssoActive: PropTypes.bool,
+    authErrors: PropTypes.arrayOf(PropTypes.object),
+    onSubmit: PropTypes.func.isRequired,
+    handleSSOLogin: PropTypes.func.isRequired,
   };
 
-  static contextType = ConnectContext;
-
-  constructor(props) {
-    super(props);
-    this.sys = require('stripes-config'); // eslint-disable-line global-require
-    this.authnUrl = this.sys.okapi.authnUrl;
-    this.okapiUrl = this.sys.okapi.url;
-    this.tenant = this.sys.okapi.tenant;
-    if (props.autoLogin && props.autoLogin.username) {
-      this.handleSubmit(props.autoLogin);
-    }
-  }
-
-  componentWillUnmount() {
-    this.props.clearAuthErrors();
-  }
-
-  handleSuccessfulLogin = () => {
-    if (matchPath(this.props.location.pathname, '/login')) {
-      this.props.history.push('/');
-    }
-  }
-
-  handleSubmit = (data) => {
-    return requestLogin({ okapi: this.sys.okapi }, this.context.store, this.tenant, data)
-      .then(this.handleSuccessfulLogin)
-      .catch(e => {
-        console.error(e); // eslint-disable-line no-console
-      });
-  }
-
-  handleSSOLogin = () => {
-    requestSSOLogin(this.okapiUrl, this.tenant);
-  }
+  static defaultProps = {
+    authErrors: [],
+    ssoActive: false,
+  };
 
   render() {
-    const { authFailure, ssoEnabled } = this.props;
-
-    const cookieMessage = navigator.cookieEnabled ?
-      '' :
-      (
-        <Row center="xs">
-          <Col xs={6}>
-            <Headline
-              size="large"
-              tag="h3"
-            >
-              <FormattedMessage id="stripes-core.title.cookieEnabled" />
-            </Headline>
-          </Col>
-        </Row>);
+    const {
+      authErrors,
+      handleSSOLogin,
+      ssoActive,
+      onSubmit,
+    } = this.props;
 
     return (
-<<<<<<< HEAD
       <Form
         onSubmit={onSubmit}
         subscription={{ values: true }}
         render={({ form, submitting, handleSubmit, submitSucceeded, values }) => {
           const { username } = values;
           const submissionStatus = submitting || submitSucceeded;
-          const buttonDisabled = submissionStatus || !(username) || !(navigator.cookieEnabled);
+          const buttonDisabled = submissionStatus || !(username);
           const buttonLabel = submissionStatus ? 'loggingIn' : 'login';
           return (
             <main>
-              <div className={styles.wrapper} style={branding?.style?.login ?? {}}>
+              <div className={styles.wrapper} style={branding.style?.login ?? {}}>
                 <div className={styles.container}>
                   <Row center="xs">
                     <Col xs={6}>
@@ -124,7 +82,6 @@ class LoginCtrl extends Component {
                           </Headline>
                         </Col>
                       </Row>
-                      { cookieMessage }
                       <div data-test-new-username-field>
                         <Row center="xs">
                           <Col xs={6}>
@@ -147,7 +104,7 @@ class LoginCtrl extends Component {
                               name="username"
                               type="text"
                               component={TextField}
-                              inputClass={styles.loginInput}
+                              inputClass={styles.input}
                               autoComplete="username"
                               autoCapitalize="none"
                               validationEnabled={false}
@@ -184,11 +141,10 @@ class LoginCtrl extends Component {
                               value=""
                               marginBottom0
                               fullWidth
-                              inputClass={styles.loginInput}
+                              inputClass={styles.input}
                               validationEnabled={false}
                               hasClearIcon={false}
                               autoComplete="current-password"
-                              required
                             />
                           </Col>
                         </Row>
@@ -200,7 +156,7 @@ class LoginCtrl extends Component {
                               buttonStyle="primary"
                               id="clickable-login"
                               type="submit"
-                              buttonClass={styles.loginSubmitButton}
+                              buttonClass={styles.submitButton}
                               disabled={buttonDisabled}
                               fullWidth
                               marginBottom0
@@ -269,12 +225,4 @@ class LoginCtrl extends Component {
   }
 }
 
-const mapStateToProps = state => ({
-  authFailure: state.okapi.authFailure,
-  ssoEnabled: state.okapi.ssoEnabled,
-});
-const mapDispatchToProps = dispatch => ({
-  clearAuthErrors: () => dispatch(setAuthError([])),
-});
-
-export default reduxConnect(mapStateToProps, mapDispatchToProps)(withRouter(LoginCtrl));
+export default LoginForm;