+++ src/loginServices.js
@@ -1,5 +1,5 @@
 import localforage from 'localforage';
-import { config, translations } from 'stripes-config';
+import { config, okapi, translations } from 'stripes-config';
 import rtlDetect from 'rtl-detect';
 import moment from 'moment';
 import { loadDayJSLocale } from '@folio/stripes-components';
@@ -51,7 +51,7 @@ export const supportedLocales = [
   'it-IT',  // italian, italy
   'ja',     // japanese
   'ko',     // korean
-  'nb',     // norwegian bokmål"
+  'nb',     // norwegian bokmål
   'nn',     // norwegian nynorsk
   'pl',     // polish
   'pt-BR',  // portuguese, brazil
@@ -358,15 +358,16 @@ export function getBindings(okapiUrl, store, tenant) {
 /**
  * loadResources
  * return a promise that retrieves the tenant's locale, user's locale,
- * plugins, and key-bindings from mod-configuration.
- * @param {string} okapiUrl
+ * plugins, and key-bindings from mod-configuration, as well as retreiving
+ * module information directly from okapi.
+ *
  * @param {redux store} store
  * @param {string} tenant
  * @param {string} userId user's UUID
  *
  * @returns {Promise}
  */
-function loadResources(okapiUrl, store, tenant, userId) {
+function loadResources(store, tenant, userId) {
   let promises = [];
 
   // tenant's locale, plugin, bindings, and user's locale are all stored
@@ -374,10 +375,10 @@ function loadResources(okapiUrl, store, tenant, userId) {
   // read-permission for configuration entries.
   if (canReadConfig(store)) {
     promises = [
-      getLocale(okapiUrl, store, tenant),
-      getUserLocale(okapiUrl, store, tenant, userId),
-      getPlugins(okapiUrl, store, tenant),
-      getBindings(okapiUrl, store, tenant),
+      getLocale(okapi.url, store, tenant),
+      getUserLocale(okapi.url, store, tenant, userId),
+      getPlugins(okapi.url, store, tenant),
+      getBindings(okapi.url, store, tenant),
     ];
   }
 
@@ -512,7 +513,6 @@ export async function logout(okapiUrl, store) {
  * Dispatch the session object, then return a Promise that fetches
  * and dispatches tenant resources.
  *
- * @param {string} okapiUrl
  * @param {object} store
  * @param {string} tenant
  * @param {string} token
@@ -520,7 +520,9 @@ export async function logout(okapiUrl, store) {
  *
  * @returns {Promise}
  */
-export function createOkapiSession(okapiUrl, store, tenant, token, data) {
+export function createOkapiSession(store, tenant, token, data) {
+  // @@ new StripesSession(store, data);
+
   // clear any auth-n errors
   store.dispatch(setAuthError(null));
 
@@ -562,7 +564,7 @@ export function createOkapiSession(okapiUrl, store, tenant, token, data) {
     .then(() => {
       store.dispatch(setIsAuthenticated(true));
       store.dispatch(setSessionData(okapiSess));
-      return loadResources(okapiUrl, store, sessionTenant, user.id);
+      return loadResources(store, sessionTenant, user.id);
     });
 }
 
@@ -652,21 +654,32 @@ export function handleLoginError(dispatch, resp) {
  * processOkapiSession
  * create a new okapi session with the response from either a username/password
  * authentication request or a bl-users/_self request.
- * @param {string} okapiUrl
+ * response body is shaped like
+ * {
+    'access_token': 'SOME_STRING',
+    'expires_in': 420,
+    'refresh_expires_in': 1800,
+    'refresh_token': 'SOME_STRING',
+    'token_type': 'Bearer',
+    'not-before-policy': 0,
+    'session_state': 'SOME_UUID',
+    'scope': 'profile email'
+ * }
+ *
  * @param {redux store} store
  * @param {string} tenant
  * @param {Response} resp HTTP response
  *
  * @returns {Promise} resolving with login response body, rejecting with, ummmmm
  */
-export function processOkapiSession(okapiUrl, store, tenant, resp, ssoToken) {
-  const token = resp.headers.get('X-Okapi-Token') || ssoToken;
+export function processOkapiSession(store, tenant, resp, ssoToken) {
   const { dispatch } = store;
 
   if (resp.ok) {
     return resp.json()
       .then(json => {
-        return createOkapiSession(okapiUrl, store, tenant, token, json)
+        const token = json.access_token || ssoToken;
+        return createOkapiSession(store, tenant, token, json)
           .then(() => json);
       })
       .then((json) => {
@@ -769,7 +782,7 @@ export function checkOkapiSession(okapiUrl, store, tenant) {
  * requestLogin
  * authenticate with a username and password. return a promise that posts the values
  * and then processes the result to begin a session.
- * @param {string} okapiUrl
+ * @param {object} okapi object from stripes.config.js
  * @param {redux store} store
  * @param {string} tenant
  * @param {object} data
@@ -777,15 +790,29 @@ export function checkOkapiSession(okapiUrl, store, tenant) {
  * @returns {Promise}
  */
 export function requestLogin(okapiUrl, store, tenant, data) {
-  const loginPath = config.useSecureTokens ? 'login-with-expiry' : 'login';
-  return fetch(`${okapiUrl}/bl-users/${loginPath}?expandPermissions=true&fullPermissions=true`, {
-    body: JSON.stringify(data),
-    credentials: 'include',
-    headers: { 'X-Okapi-Tenant': tenant, 'Content-Type': 'application/json' },
-    method: 'POST',
-    mode: 'cors',
-  })
-    .then(resp => processOkapiSession(okapiUrl, store, tenant, resp));
+  // got Keycloak?
+  if (okapi.authnUrl) {
+    return fetch(okapi.authnUrl, {
+      method: 'POST',
+      body: new URLSearchParams({
+        ...data,
+        'grant_type': 'password',
+        'client_id': okapi.clientId,
+        'client_secret': okapi.clientSecret,
+      })
+    })
+      .then(resp => processOkapiSession(store, tenant, resp));
+  } else {
+    // legacy built-in authentication
+    const loginPath = config.useSecureTokens ? 'login-with-expiry' : 'login';
+    return fetch(`${okapiUrl}/bl-users/${loginPath}?expandPermissions=true&fullPermissions=true`, {
+      body: JSON.stringify(data),
+      credentials: 'include',
+      headers: { 'X-Okapi-Tenant': tenant, 'Content-Type': 'application/json' },
+      method: 'POST',
+      mode: 'cors',
+    })
+    .then(resp => processOkapiSession(store, tenant, resp));
 }
 
 /**
@@ -799,8 +826,9 @@ export function requestLogin(okapiUrl, store, tenant, data) {
  * @returns {Promise} Promise resolving to the response of the request
  */
 function fetchUserWithPerms(okapiUrl, tenant, token, rtrIgnore = false) {
+  const usersPath = okapi.authnUrl ? 'users-keycloak' : 'bl-users';
   return fetch(
-    `${okapiUrl}/bl-users/_self?expandPermissions=true&fullPermissions=true`,
+    `${okapiUrl}/${usersPath}/_self?expandPermissions=true&fullPermissions=true`,
     {
       headers: getHeaders(tenant, token),
       rtrIgnore,
@@ -819,7 +847,15 @@ function fetchUserWithPerms(okapiUrl, tenant, token, rtrIgnore = false) {
  */
 export function requestUserWithPerms(okapiUrl, store, tenant, token) {
   return fetchUserWithPerms(okapiUrl, tenant, token, !token)
-    .then(resp => processOkapiSession(okapiUrl, store, tenant, resp, token));
+    .then((resp) => {
+      if (resp.ok) {
+        return processOkapiSession(store, tenant, resp, token);
+      } else {
+        return resp.json().then((error) => {
+          throw error;
+        });
+      }
+    });
 }
 
 /**
@@ -869,7 +905,7 @@ export function updateUser(store, data) {
  * updateTenant
  * 1. prepare user info for requested tenant
  * 2. update okapi session
- * @param {object} okapi
+ * @param {object} okapiConfig
  * @param {string} tenant
  *
  * @returns {Promise}