+++ src/okapiActions.js
@@ -135,6 +135,13 @@ function updateCurrentUser(data) {
   };
 }
 
+function setOkapiTenant(payload) {
+  return {
+    type: 'SET_OKAPI_TENANT',
+    payload
+  };
+}
+
 function setTokenExpiration(tokenExpiration) {
   return {
     type: 'SET_TOKEN_EXPIRATION',
@@ -187,4 +194,5 @@ export {
   setTranslations,
   toggleRtrModal,
   updateCurrentUser,
+  setOkapiTenant
 };